# back-end/swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "Node JS API project",
      "version": "1.0.0"
    },
    "servers": [
      {
        "url": "http://localhost:4000/"
      }
    ],
    "paths": {
...
      "/data_sensor/addData": {
        "post": {
          "summary": "Add new sensor data",
          "description": "Add new sensor data to the database",
          "tags": ["Data Sensor"],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "temperature": {
                      "type": "number"
                    },
                    "humidity": {
                      "type": "number"
                    },
                    "light": {
                      "type": "number"
                    }
                  }
                }
              }
            }
          },
          "responses": {
            "201": {
              "description": "Data added successfully"
            },
            "400": {
              "description": "Bad request. Missing required fields or invalid data"
            },
            "500": {
              "description": "Internal Server Error"
            }
          }
        }
      },
      "/data_sensor/{id}": {
        "put": {
          "summary": "Update sensor data by ID",
          "description": "Update sensor data in the database by ID",
          "tags": ["Data Sensor"],
          "parameters": [
            {
              "in": "path",
              "name": "id",
              "required": true,
              "description": "Numeric ID of the sensor data to update",
              "schema": {
                "type": "integer"
              }
            }
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "temperature": {
                      "type": "number"
                    },
                    "humidity": {
                      "type": "number"
                    },
                    "light": {
                      "type": "number"
                    }
                  }
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Data updated successfully"
            },
            "400": {
              "description": "Bad request. Missing required fields or invalid data"
            },
            "404": {
              "description": "Data not found"
            },
            "500": {
              "description": "Internal Server Error"
            }
          }
        },
        "delete": {
          "summary": "Delete sensor data by ID",
          "description": "Delete sensor data from the database by ID",
          "tags": ["Data Sensor"],
          "parameters": [
            {
              "in": "path",
              "name": "id",
              "required": true,
              "description": "Numeric ID of the sensor data to delete",
              "schema": {
                "type": "integer"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Data deleted successfully"
            },
            "404": {
              "description": "Data not found"
            },
            "500": {
              "description": "Internal Server Error"
            }
          }
        }
      }     ,
...
      "/action_history/addData": {
        "post": {
          "summary": "Add new action history data",
          "description": "Add new action history data to the database",
          "tags": ["Action History"],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "Device": {
                      "type": "string"
                    },
                    "action": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "responses": {
            "201": {
              "description": "Data added successfully"
            },
            "400": {
              "description": "Bad request. Missing required fields or invalid data"
            },
            "500": {
              "description": "Internal Server Error"
            }
          }
        }
      },
      "/action_history/{id}": {
        "put": {
            "summary": "Update action history data by ID",
            "description": "Update action history data in the database by specifying the ID",
            "tags": ["Action History"],
            "parameters": [
                {
                    "name": "id",
                    "in": "path",
                    "description": "ID of the action history data to update",
                    "required": true,
                    "schema": {
                        "type": "integer"
                    }
                }
            ],
            "requestBody": {
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "device": {
                                    "type": "string"
                                },
                                "action": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            },
            "responses": {
                "200": {
                    "description": "Data updated successfully"
                },
                "400": {
                    "description": "Bad request. Missing required fields or invalid data"
                },
                "404": {
                    "description": "Data not found"
                },
                "500": {
                    "description": "Internal server error"
                }
            }
        },
        "delete": {
          "summary": "Delete action history data by ID",
          "description": "Delete action history data from the database by specifying the ID",
          "tags": ["Action History"],
          "parameters": [
              {
                  "name": "id",
                  "in": "path",
                  "description": "ID of the action history data to delete",
                  "required": true,
                  "schema": {
                      "type": "integer"
                  }
              }
          ],
          "responses": {
              "200": {
                  "description": "Data deleted successfully"
              },
              "404": {
                  "description": "Data not found"
              },
              "500": {
                  "description": "Internal server error"
              }
          }
      }
      
    }
    }      
}